Input protection for add, delete and read

+++ OODataBase/DataBase/DeleteItems/DeleteItem.xaml.cs
@@ -31,26 +31,46 @@ namespace DataBase.DeleteItems
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            if (DB.Delete(comboBox_Items.SelectedItem.ToString(), Convert.ToInt32(id.Text)))
+            if (comboBox_Items.SelectedItem == null || id.Text == "")
             {
-                MessageBoxResult result = MessageBox.Show("Deleted successfully",
-                                          "Information",
-                                          MessageBoxButton.OK,
-                                          MessageBoxImage.Information);
-                if (result == MessageBoxResult.OK)
-                {
-                    this.Close();
-                }
-            }
-            else
-            {
-                MessageBoxResult result = MessageBox.Show("Unable to delete item!",
+                MessageBoxResult result = MessageBox.Show("Type and ID are required!",
                                           "Information",
                                           MessageBoxButton.OK,
                                           MessageBoxImage.Error);
-                if (result == MessageBoxResult.OK)
+            }
+            else
+            {
+                if (!Int32.TryParse(id.Text, out int tmp))
                 {
-                    id.Text = "";
+                    MessageBoxResult result = MessageBox.Show("ID must be a number!",
+                                          "Information",
+                                          MessageBoxButton.OK,
+                                          MessageBoxImage.Error);
+                }
+                else
+                {
+                    if (DB.Delete(comboBox_Items.SelectedItem.ToString(), Convert.ToInt32(id.Text)))
+                    {
+                        MessageBoxResult result = MessageBox.Show("Deleted successfully",
+                                                  "Information",
+                                                  MessageBoxButton.OK,
+                                                  MessageBoxImage.Information);
+                        if (result == MessageBoxResult.OK)
+                        {
+                            this.Close();
+                        }
+                    }
+                    else
+                    {
+                        MessageBoxResult result = MessageBox.Show("Unable to delete item!",
+                                                  "Information",
+                                                  MessageBoxButton.OK,
+                                                  MessageBoxImage.Error);
+                        if (result == MessageBoxResult.OK)
+                        {
+                            id.Text = "";
+                        }
+                    }
                 }
             }
         }
